requirements.txt
Django==3.2.8
djangorestframework==3.12.4
drf_writable_nested==0.6.3
Pillow==8.4.0
django-filter==21.1
django-cleanup==5.2.0
django-cors-headers==3.10.0